refactor(entity preparation): moved all entity preparation flow/stream into seperate module
